app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.0.0
boto3>=1.35.0
